requirements.txt
adjustText==1.3.0
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.1.0
auto-py-to-exe==2.46.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
bottle==0.13.2
bottle-websocket==0.2.9
cachetools==5.5.2
ccxt==4.4.82
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==45.0.2
curl_cffi==0.11.1
cycler==0.12.1
Cython==3.0.11
dash==3.0.4
dash-bootstrap-components==2.0.2
dask==2025.1.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Eel==0.18.1
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.0.3
fonttools==4.55.6
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2024.12.0
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
gitdb==4.0.12
GitPython==3.1.44
gluonts==0.15.1
greenlet==3.1.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.6
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.14
kaleido==0.2.1
kiwisolver==1.4.8
lightning==2.1.4
lightning-utilities==0.14.3
locket==1.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
missingno==0.5.2
mistune==3.1.2
mlxtend==0.23.4
mpmath==1.3.0
multidict==6.4.3
multitasking==0.0.11
narwhals==1.33.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.24.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
patsy==1.0.1
peewee==3.18.1
pefile==2023.2.7
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
pmdarima==2.0.4
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil==6.1.1
pure_eval==0.2.3
pyarray==0.2
pyarrow==19.0.0
pycares==4.8.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.0.1
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
soupsieve==2.6
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.44.1
sympy==1.14.0
ta==0.11.0
tenacity==9.0.0
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
toolz==0.12.1
torch==2.0.1
torchmetrics==1.7.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.0.6
widgetsnbextension==4.0.14
yarl==1.20.0
yfinance==0.2.61
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
